refactor: unify Operation targets on M (u16)
Every Operation variant now carries its target as M instead of
mixing N (u8) for Add/Subtract/Divide/Given with M (u16) for
Multiply. Multiply needs the wider type for products, the others
fit in N for any practical KenKen but use M so consumers reading
a target value don't have to match on the variant first.

- Update M's docstring to reflect its broader role.
- arithmetic::cage_tuples: introduce a small `fits_in_n` helper
  that narrows targets back to N for the additive helpers, which
  keep their N math.
- constraints::valid_targets / is_valid: ranges and comparisons
  in M.
- generation::default_op_policy and the puzzle.rs test helpers:
  M::from on N-typed inputs.
- public_api.rs custom_op_policy: same.
- Test helpers consolidated into a single collect_targets that
  works for any operator now that all targets share a type.

+++ src/puzzle.rs
@@ -400,7 +400,7 @@ mod tests {
             .iter()
             .map(|&(row, column)| Cell::new(row, column))
             .collect();
-        Cage::new(n, Polyomino::new(&cells), Operation::Add(n))
+        Cage::new(n, Polyomino::new(&cells), Operation::Add(u16::from(n)))
     }
 
     fn poly(cells: &[(usize, usize)]) -> Polyomino {
@@ -571,12 +571,16 @@ mod tests {
 
     fn given(row: usize, column: usize, value: u8) -> Cage {
         let cell = Cell::new(row, column);
-        Cage::new(value, Polyomino::new(&[cell]), Operation::Given(value))
+        Cage::new(
+            value,
+            Polyomino::new(&[cell]),
+            Operation::Given(u16::from(value)),
+        )
     }
 
     fn row_add(n: u8, row: usize, target: u8) -> Cage {
         let cells: Vec<Cell> = (0..n as usize).map(|col| Cell::new(row, col)).collect();
-        Cage::new(n, Polyomino::new(&cells), Operation::Add(target))
+        Cage::new(n, Polyomino::new(&cells), Operation::Add(u16::from(target)))
     }
 
     #[test]
@@ -799,7 +803,7 @@ mod tests {
             .iter()
             .map(|&(row, column)| Cell::new(row, column))
             .collect();
-        Cage::new(n, Polyomino::new(&cells), Operation::Add(target))
+        Cage::new(n, Polyomino::new(&cells), Operation::Add(u16::from(target)))
     }
 
     #[test]
@@ -957,7 +961,11 @@ mod tests {
             .iter()
             .map(|&(row, column)| Cell::new(row, column))
             .collect();
-        Cage::new(n, Polyomino::new(&cells), Operation::Subtract(target))
+        Cage::new(
+            n,
+            Polyomino::new(&cells),
+            Operation::Subtract(u16::from(target)),
+        )
     }
 
     #[test]
